Coerce to a "raster" object

Implementation of the generic as.raster function to create a "raster" (small r) object. Such objects can be used for plotting with the rasterImage function. NOT TO BE CONFUSED with the Raster* (big R) objects defined by the 'raster' package!

# S4 method for SpatRaster
as.raster(x, maxcell=500000, col)

Arguments

x

SpatRaster

maxcell

positive integer. Maximum number of cells to use for the plot

col

vector of colors. Default is col=rev(terrain.colors(255)))

Value

'raster' object

Examples

r <- rast(ncols=3, nrows=3)
values(r) <- 1:ncell(r)
as.raster(r)
#>      [,1]      [,2]      [,3]     
#> [1,] "#F2F2F2" "#EFC2B3" "#ECB178"
#> [2,] "#E9BC3C" "#E6E402" "#A1D600"
#> [3,] "#63C600" "#2EB600" "#00A600"