Disaggregate raster cells

SpatRaster: Create a SpatRaster with a higher resolution (smaller cells). The values in the new SpatRaster are the same as in the larger original cells.

SpatVector: Separate multi-objects (points, lines, polygons) into single objects.

# S4 method for SpatRaster
disaggregate(x, fact, method="near", filename="", ...)

# S4 method for SpatVector
disaggregate(x)

Arguments

x
SpatRaster or SpatVector

fact
positive integer. Aggregation factor expressed as number of cells in each direction (horizontally and vertically). Or two integers (horizontal and vertical aggregation factor) or three integers (when also aggregating over layers)

method
character. Either "near" for nearest or "bilinear" for bilinear interpolation

filename
character. Output filename

...
additional arguments for writing files as in writeRaster

See also

Value

SpatRaster

Examples

r <- rast(ncols=10, nrows=10)
rd <- disaggregate(r, fact=c(10, 2))
ncol(rd)

#> [1] 20

nrow(rd)

#> [1] 100

values(r) <- 1:ncell(r)
rd <- disaggregate(r, fact=c(4, 2))