Extend

Enlarge the spatial extent of a SpatRaster. See crop if you (also) want to remove rows or columns.

You can also enlarge a SpatExtent with this method, or with algebraic notation (see examples)

# S4 method for SpatRaster
extend(x, y, filename="", overwrite=FALSE, ...) 

# S4 method for SpatExtent
extend(x, y)

Arguments

x
SpatRaster or SpatExtent

y
If x is a SpatRaster, y should be a SpatExtent, or an object from which it can be extracted (such as SpatRaster and SpatVector objects). Alternatively, you can provide two positive integers indicating the number of rows and columns that need to be added at each side (or a single positive integer when the number of rows and columns is equal)

If x is a SpatExtent, y should be a numeric vector of 1, 2, or 4 elements

filename
character. Output filename

overwrite
logical. If TRUE, filename is overwritten

...
additional arguments for writing files as in writeRaster

Value

SpatRaster or SpatExtent

See also

Examples

r <- rast(xmin=-150, xmax=-120, ymin=30, ymax=60, ncols=36, nrows=18)
values(r) <- 1:ncell(r)
e <- ext(-180, -100, 40, 70)
re <- extend(r, e)

# expand with a number of rows and columns (at each side)
re2 <- extend(r, c(2,10))

# SpatExtent
e <- ext(r)
e

#> SpatExtent : -150, -120, 30, 60 (xmin, xmax, ymin, ymax)

extend(e, 10)

#> SpatExtent : -160, -110, 20, 70 (xmin, xmax, ymin, ymax)

extend(e, c(10, -10, 0, 20))

#> SpatExtent : -160, -130, 30, 80 (xmin, xmax, ymin, ymax)