global statistics

Compute global statistics, that is summarized values of an entire SpatRaster.

If x is very large global will fail, except when fun is one of "mean", "min", "max", "sum", "prod", "range" (min and max), "rms" (root mean square), "sd" (sample standard deviation), "sdpop" (population standard deviation), "isNA" (number of cells that are NA), "notNA" (number of cells that are not NA).

You can compute a weighted mean or sum by providing a SpatRaster with weights.

# S4 method for SpatRaster
global(x, fun="mean", weights=NULL, ...)

Arguments

x

SpatRaster

fun

function to be applied to summarize the values by zone. Either as one of these character values: "max", "min", "mean", "sum", "range", "rms" (root mean square), "sd", "std" (population sd, using n rather than n-1), "isNA", "notNA"; or, for relatively small SpatRasters, a proper function

...

additional arguments passed on to fun

weights

NULL or SpatRaster

Value

A data.frame with a row for each layer

See also

Examples

r <- rast(ncols=10, nrows=10)
values(r) <- 1:ncell(r)
global(r, "sum")
#>        sum
#> lyr.1 5050
global(r, "mean", na.rm=TRUE)
#>       mean
#> lyr.1 50.5