Intersection

Intersect the geometries of two SpatVectors.

Intersecting points with points uses the extent of y to get the intersection. Intersecting of points and lines is not supported because of numerical inaccuracies with that. You can use buffer, to create polygons from lines and use these with intersect.

See crop for intersection of a SpatRaster.

# S4 method for SpatVector,SpatVector
intersect(x, y)

# S4 method for SpatVector,SpatExtent
intersect(x, y)

# S4 method for SpatExtent,SpatVector
intersect(x, y)

# S4 method for SpatExtent,SpatExtent
intersect(x, y)

Arguments

x
SpatVector or SpatExtent

y
SpatVector or SpatExtent

Value

Same as x

See also

Examples

e1 <- ext(-10, 10, -20, 20)
e2 <- ext(0, 20, -40, 5)
intersect(e1, e2)

#> SpatExtent : 0, 10, -20, 5 (xmin, xmax, ymin, ymax)

f <- system.file("ex/lux.shp", package="terra")
v <- vect(f)
e <- ext(5.6, 6, 49.55, 49.7)
x <- intersect(v, e)

p <- vect(c("POLYGON ((5.8 49.8, 6 49.9, 6.15 49.8, 6 49.6, 5.8 49.8))", 
"POLYGON ((6.3 49.9, 6.2 49.7, 6.3 49.6, 6.5 49.8, 6.3 49.9))"), crs=crs(v))
values(p) <- data.frame(pid=1:2, area=expanse(p))

y <- intersect(v, p)