Arithmetic, logical and general mathematical methods

Standard operators and mathematical methods for computations with SpatRaster objects. Computations are local (applied on a cell by cell basis). If multiple SpatRaster objects are used, these must have the same extent and resolution. These have been implemented:

Arith: +, -, *, /, ^, %%, %/%

Compare: ==, !=, >, <, <=, >=, is.na, is.nan, is.finite, is.infinite

The terra package does not distinguish between NA (not available) and NaN (not a number). In most cases this state is represented by NaN.

Logical: !, &, |, isTRUE, isFALSE

Summary: max, min, range, prod, sum, any, all

Math: abs, sign, sqrt, ceiling, floor, trunc, cummax, cummin, cumprod, cumsum, log, log10, log2, log1p, acos, acosh, asin, asinh, atan, atanh, exp, expm1, cos, cosh, sin, sinh, tan, tanh, round, signif

The following methods have been implemented for SpatExtent: round, floor, ceil, ==, for (SpatExtent, SpatExtent): ==, +, -, and for (SpatExtent, numeric) : +, -, *, /, %%

# S4 method for SpatRaster
log(x, base=exp(1))

Arguments

x
SpatRaster

base
a positive or complex number: the base with respect to which logarithms are computed

seealso

ifel to convieniently combine operations and app to use mathematical functions not implemented by the package.

Value

SpatRaster or SpatExtent

Examples

r1 <- rast(ncols=10, nrows=10)
v <- runif(ncell(r1))
v[10:20] <- NA
values(r1) <- v
r2 <- rast(r1)
values(r2) <- 1:ncell(r2) / ncell(r2)
r3 <- r1 + r2
r2 <- r1 / 10
r3 <- r1 * (r2 - 1 / r2)

b <- c(r1, r2, r3)
b2 <- b * 10
s <- sqrt(b2)
round(s, 1)

#> class       : SpatRaster 
#> dimensions  : 10, 10, 3  (nrow, ncol, nlyr)
#> resolution  : 36, 18  (x, y)
#> extent      : -180, 180, -90, 90  (xmin, xmax, ymin, ymax)
#> coord. ref. : lon/lat WGS 84 
#> source      : memory 
#> names       : lyr.1, lyr.1, lyr.1 
#> min values  :   0.8,   0.2,   NaN 
#> max values  :   3.1,   1.0,   NaN 

max(s)

#> class       : SpatRaster 
#> dimensions  : 10, 10, 1  (nrow, ncol, nlyr)
#> resolution  : 36, 18  (x, y)
#> extent      : -180, 180, -90, 90  (xmin, xmax, ymin, ymax)
#> coord. ref. : lon/lat WGS 84 
#> source      : memory 
#> name        : max 
#> min value   : NaN 
#> max value   : NaN 

max(s, na.rm=TRUE)

#> class       : SpatRaster 
#> dimensions  : 10, 10, 1  (nrow, ncol, nlyr)
#> resolution  : 36, 18  (x, y)
#> extent      : -180, 180, -90, 90  (xmin, xmax, ymin, ymax)
#> coord. ref. : lon/lat WGS 84 
#> source      : memory 
#> name        :       max 
#> min value   : 0.7705493 
#> max value   :  3.106089 

x <- is.na(s)

y <- which.max(s)

### SpatExtent methods
x <- ext(0.1, 2.2, 0, 3)
y <- ext(-2, 1, -2,2)
# union
x + y

#> SpatExtent : -2, 2.2, -2, 3 (xmin, xmax, ymin, ymax)

# intersection
x * y

#> SpatExtent : 0.1, 1, 0, 2 (xmin, xmax, ymin, ymax)

e <- x 
e

#> SpatExtent : 0.1, 2.2, 0, 3 (xmin, xmax, ymin, ymax)

e * 2

#> SpatExtent : -0.95, 3.25, -1.5, 4.5 (xmin, xmax, ymin, ymax)

e / 2

#> SpatExtent : 0.625, 1.675, 0.75, 2.25 (xmin, xmax, ymin, ymax)

e + 1

#> SpatExtent : -0.9, 3.2, -1, 4 (xmin, xmax, ymin, ymax)

e - 1

#> SpatExtent : 1.1, 1.2, 1, 2 (xmin, xmax, ymin, ymax)