Make a map

Plot the values of a SpatRaster or SpatVector to make a map. See lines to add a SpatVector to an existing map.

# S4 method for SpatRaster,numeric
plot(x, y=1, col, type, mar=NULL, legend=TRUE, axes=TRUE, plg=list(), pax=list(), 
    maxcell=500000, smooth=FALSE, range=NULL, levels=NULL, all_levels=FALSE,
  breaks=NULL, breakby="eqint", fun=NULL, colNA=NULL, alpha=NULL, sort=FALSE, 
  decreasing=FALSE, grid=FALSE, ext=NULL, reset=FALSE, ...)

# S4 method for SpatRaster,missing
plot(x, y, maxcell=500000, main, mar=NULL, nc, nr, maxnl=16, legend, ...)

# S4 method for SpatRaster,character
plot(x, y, ...)

# S4 method for SpatVector,character
plot(x, y, col, type, mar=NULL, legend=TRUE, add=FALSE, axes=!add, 
    main=y, buffer=TRUE, background=NULL, grid=FALSE, ext=NULL, 
    sort=TRUE, decreasing=FALSE, plg=list(), pax=list(), nr, nc, ...)

# S4 method for SpatVector,numeric
plot(x, y, ...)

# S4 method for SpatVector,missing
plot(x, y, values=NULL, ...)

# S4 method for SpatExtent,missing
plot(x, y, ...)

# S4 method for SpatGraticule,missing
plot(x, y, background=NULL, col="black", mar=c(1,1,1,1), 
  labels=TRUE, retro=FALSE, labloc=c(1,1), inclon=NULL, inclat=NULL, ...)

Arguments

x

SpatRaster or SpatVector

y

missing or positive integer or name indicating the layer(s) to be plotted

col

character. Colors. The default is rev(grDevices::terrain.colors(50)). If x is a SpatRaster, it can also be a data.frame with two columns (value, color) to get a "classes" type legend or with three columns (from, to, color) to get an "interval" type legend

type

character. Type of map/legend. One of "continuous", "classes", or "interval". If not specified, the type is chosen based on the data

mar

numeric vector of length 4 to set the margins of the plot (to make space for the legend). The default is (3.1, 3.1, 2.1, 7.1) for a single plot with a legend and (3.1, 3.1, 2.1, 2.1) otherwise. Use mar=NA to not set the margins

legend

logical or character. If not FALSE a legend is drawn. The character value can be used to indicate where the legend is to be drawn. For example "topright" or "bottomleft". Use plg for more refined placement. Not supported for continuous legends (the default for raster data)

axes

logical. Draw axes?

buffer

logical. If TRUE the plotting area is slightly larger than the extent of x

background

background color. Default is no color (white)

plg

list with parameters for drawing the legend. See the arguments for legend. Argument digits can be used to set the number of digits to print (after the decimal point for a continuous legend)

pax

list with parameters for drawing axes. See the arguments for axis. Arguments side, tick and lab can be used to indicate for which of the four axes to draw a line (side), tick mark, or label. The default is c(1:4) for side and 1:2 for the other two. If side is changed the other two default to that value. Logical argument retro can be used to use a sexagesimal notation for the labels (degrees/minutes/hemisphere) instead of the standard decimal notation

maxcell

positive integer. Maximum number of cells to use for the plot

smooth

logical. If TRUE the cell values are smoothed (only if a continuous legend is used)

range

numeric. minimum and maximum values to be used for the continuous legend

levels

character. labels for the legend when type="classes"

all_levels

logical. If TRUE, the legend shows all levels of a categorical raster, even if they are not present in the data

breaks

numeric. Either a single number to indicate the number of breaks desired, or the actual breaks. When providing this argument, the default legend becomes "interval"

breakby

character or function. Either "eqint" for equal interval breaks, "cases" for equal quantile breaks. If a function is supplied it should take a single argument (a vector of values) and create groups

fun

function to be called after plotting each SpatRaster layer to add something to each map (such as text, legend, lines). For example, with SpatVector v, you could do fun=function() lines(v). The function may have one argument, representing the the layer that is plotted (1 to the number of layers)

colNA

character. color for the NA values

alpha

Either a single numeric between 0 and 1 to set the transparency for all colors (0 is transparent, 1 is opaque) or a SpatRaster with values between 0 and 1 to set the transparency by cell. To set the transparency for a given color, set it to the colors directly

sort

logical. If TRUE legends with categorical values are sorted. If x is a SpatVector you can also supply a vector of the unique values, in the order in which you want them to appear in the legend

decreasing

logical. If TRUE, legends are sorted in decreasing order

grid

logical. If TRUE grid lines are drawn. Their properties such as type and color can be set with the pax argument

nc

positive integer. Optional. The number of columns to divide the plotting device in (when plotting multiple layers)

nr

positive integer. Optional. The number of rows to divide the plotting device in (when plotting multiple layers)

main

character. Main plot titles (one for each layer to be plotted)

maxnl

positive integer. Maximum number of layers to plot (for a multi-layer object)

add

logical. If TRUE add the object to the current plot

ext

SpatExtent. Can be use instead of xlim and ylim to set the extent of the plot

reset

logical. If TRUE add the margins (see argument mar) are reset to what they were before calling plot; doing so may affect the display of additional objects that are added to the map (e.g. with lines

values

Either a vector with values to be used for plotting or a two-column data.frame, where the first column matches a variable in x and the second column has the values to be plotted

...

arguments passed to plot("SpatRaster", "numeric") and/or additional graphical arguments

labels

logical. If TRUE, show graticule labels

retro

logical. If TRUE, show "retro" instead of decimal labels with the graticule

labloc

numeric. The first number indicates where the longitude graticule labels should be drawn (1=bottom, 2=top, NA=not drawn, any other number=top and bottom). The second number indicates where the latitude graticule labels should be drawn (1=left, 2=right, NA=not drawn, any other number=left and right) inclonpositive integers between 1 and the number of labels, indicating which longitude graticule labels should be included inclatpositive integers between 1 and the number of labels, indicating which latitude graticule labels should be included

points, lines, polys, image, scatterplot, scale bar: sbar, north arrow: north, graticule: graticule

## raster f <- system.file("ex/elev.tif", package="terra") r <- rast(f) plot(r)plot(r, type="interval")e <- c(6.37, 6.41, 49.9, 50.1) plot(r, plg=list(ext=e, title="Title", title.cex=1.25), pax=list(side=1:4, retro=TRUE)) north("topleft")d <- classify(r, c(100,200,300,400,500,600)) plot(d, type="classes")plot(d, type="interval", breaks=1:5) plot(d, type="interval", breaks=c(1,4,5), plg=list(legend=c("1-4", "4-5"))) plot(d, type="classes", plg=list(legend=c("Mr", "Xx", "As", "Zx", "Bb"), x="bottomright"))x <- trunc(r/200) levels(x) <- data.frame(id=0:2, element=c("earth", "wind", "fire")) plot(x, plg=list(x="topright"),mar=c(2,2,2,2))# two plots with the same legend dev.new(width=6, height=4, noRStudioGD = TRUE) par(mfrow=c(1,2)) plot(r, range=c(50,600)) plot(r/2, range=c(50,600))# as you only need one legend: par(mfrow=c(1,2)) plot(r, range=c(50,600), mar=c(4, 3, 4, 3), plg=list(shrink=0.9, cex=.8), pax=list(side=1:2, cex.axis=.6)) #text(182500, 335000, "Two maps, one plot", xpd=NA) plot(r/2, range=c(50,600), mar=c(4, 2, 4, 4), legend=FALSE, pax=list(side=c(1,4), cex.axis=.6))# multi-layer with RGB s <- rast(system.file("ex/logo.tif", package="terra")) s plot(s) # remove RGB plot(s*1) # or use layers plot(s, 1) plot(s, 1:3)# fix legend by linking values and colorsx = rast(nrows = 2, ncols = 2, vals=1) y = rast(nrows = 2, ncols = 2, vals=c(1,2,2,1)) cols = data.frame(id=1:2, col=c("red", "blue")) plot(c(x,y), col=cols)r = rast(nrows=10, ncols=10, vals=1:100) dr = data.frame(from=c(5,33,66,150), to=c(33, 66, 95,200), col=rainbow(4)) plot(r, col=dr)### SpatVector f <- system.file("ex/lux.shp", package="terra") v <- vect(f)plot(v)plot(v, 2, pax=list(side=1:2), plg=list(x=6.2, y=50.2, cex=1.2))plot(v, 4, pax=list(side=1:2), plg=list(x=6.2, y=50.2, ncol=2), main="")plot(v, 1, plg=list(x=5.9, y=49.37, horiz=TRUE, cex=1.1), main="", mar=c(5,2,0.5,0.5))plot(v, density=1:12, angle=seq(18, 360, 20), col=rainbow(12))plot(v, "NAME_2", col=rainbow(12), border=c("gray", "blue"), lwd=3, type="classes")plot(v, "AREA", type="interval", breaks=3, mar=c(3.1, 3.1, 2.1, 3.1), plg=list(x="topright"), main="")plot(v, "AREA", type="interval", breaks=c(0,200,250,350), mar=c(2,2,2,2), plg=list(legend=c("<200", "200-250", ">250"), cex=1, bty="o", x=6.4, y=50.125, box.lwd=2, bg="light yellow", title="My Legend"))

methods

spatial