Take a regular sample

Take a spatial sample from a SpatRaster, SpatVector or SpatExtent. Sampling a SpatVector or SpatExtent always returns a SpatVector of points.

With a SpatRaster, you can get cell values, cell numbers (cells=TRUE), coordinates (xy=TRUE) or (when type="regular" and as.raster=TRUE) get a new SpatRaster with the same extent, but fewer cells.

In order to assure regularity when requesting a regular sample, the number of cells or points returned may not be exactly the same as the size requested.

# S4 method for SpatRaster
spatSample(x, size, method="random", replace=FALSE, na.rm=FALSE, as.raster=FALSE,
  as.points=FALSE, values=TRUE, cells=FALSE, xy=FALSE, ext=NULL, warn=TRUE)

# S4 method for SpatVector
spatSample(x, size, method="random", strata=NULL, chess="")

# S4 method for SpatExtent
spatSample(x, size, method="random", lonlat)

Arguments

x
SpatRaster

size
numeric. The sample size. If x is a SpatVector, you can also provide a vector of the same length as x in wich case sampling is done seperately for each geometry

method
character. Should be "regular" or "random". It can also be "stratified" if x is a SpatVector

replace
logical. If TRUE, sampling is with replacement (if method="random"

na.rm
logical. If TRUE, codeNAs are removed. Only used with random sampling of cell values. That is with method="random", as.raster=FALSE, cells=FALSE

as.raster
logical. If TRUE, a SpatRaster is returned

as.points
logical. If TRUE, a SpatVector of points is returned

values
logical. If TRUE cell values are returned

cells
logical. If TRUE, cell numbers are returned

xy
logical. If TRUE, cell coordinates are returned

ext
SpatExtent or NULL to restrict sampling to a a subset of the area of x

warn
logical. Give a warning if the sample size returned is smaller than requested

strata
if not NULL, stratified random sampling is done, taking size samples from each stratum. If x has polygon geometry, strata must be a field name (or index) in x. If x has point geometry, strata can be a SpatVector of polygons or a SpatRaster

chess
character. One of "", "white", or "black". For stratified sampling if strata is a SpatRaster. If not "", samples are only taken from alternate cells, organized like the "white" or "black" fields on a chessboard

lonlat
logical. If TRUE, sampling of a SpatExtent is weighted by cos(latitude). For SpatRaster and SpatVector this done based on the crs, but it is ignored if as.raster=TRUE

Value

numeric or SpatRaster

Examples

f <- system.file("ex/elev.tif", package="terra")
r <- rast(f)
s <- spatSample(r, 10, as.raster=TRUE)
spatSample(r, 10)

#>    elevation
#> 1        338
#> 2        380
#> 3        292
#> 4        351
#> 5        188
#> 6        NaN
#> 7        NaN
#> 8        497
#> 9        358
#> 10       NaN

spatSample(r, 10, "random")

#>    elevation
#> 1        371
#> 2        449
#> 3        336
#> 4        466
#> 5        368
#> 6        484
#> 7        358
#> 8        386
#> 9        345
#> 10       406

## if you require cell numbers and/or coordinates
size <- 6
# random cells 
cells <- spatSample(r, 6, "random", cells=TRUE, values=FALSE)
cells <- as.vector(cells)

v <- r[cells]
xy <- xyFromCell(r, cells)
cbind(xy, v)

#>          x        y elevation
#> 1 6.504167 50.02917       NaN
#> 2 6.220833 49.69583       402
#> 3 6.187500 49.61250       337
#> 4 5.870833 49.84583       509
#> 5 5.862500 50.11250       NaN
#> 6 6.245833 50.08750       NaN

# regular 
cells <- spatSample(r, 6, "regular", cells=TRUE, values=FALSE)
cells <- as.vector(cells)
v <- r[cells]
xy <- xyFromCell(r, cells)
cbind(xy, v)

#>          x        y elevation
#> 1 6.137500 50.12083       NaN
#> 2 6.137500 49.97083       412
#> 3 6.129167 49.82083       328
#> 4 6.129167 49.67083       228
#> 5 6.129167 49.52083       275

## SpatExtent 
e <- ext(r)
spatSample(e, 10, "random", lonlat=TRUE)

#>              x        y
#>  [1,] 6.465837 49.46577
#>  [2,] 6.269257 49.56902
#>  [3,] 6.069760 49.49101
#>  [4,] 5.887555 49.64056
#>  [5,] 6.481434 50.05977
#>  [6,] 6.513902 49.81226
#>  [7,] 5.904535 49.63760
#>  [8,] 6.153146 49.46480
#>  [9,] 5.926180 49.86753
#> [10,] 6.450160 49.65623

## SpatVector
f <- system.file("ex/lux.shp", package="terra")
v <- vect(f)
#sample geometries 
i <- sample(nrow(v), 5)
vv <- v[i,]