Apply a terra function that takes only a single layer and returns a SpatRaster to all layers of a SpatRaster

Apply to all layers of a SpatRaster a function that only takes a single layer SpatRaster and returns a SpatRaster (these are rare).

In most cases you can also use lapply or sapply for this.

# S4 method for SpatRaster
sapp(x, fun, ..., filename="", overwrite=FALSE, wopt=list())

Arguments

x

SpatRaster

fun

a function that takes a SpatRaster argument and can be applied to each layer of x

...

additional arguments to be passed to fun

filename

character. Output filename

overwrite

logical. If TRUE, filename is overwritten

wopt

list with named options for writing files as in writeRaster

Value

SpatRaster

See also

Examples

s <- rast(system.file("ex/logo.tif", package="terra")) + 1  
x <- sapp(s, terrain)