Set a window

Experimental: Assign a window (area of interest) to a SpatRaster with a SpatExtent, or set it to NULL to remove the window. This is similar to crop without actually creating a new dataset.

Currently, the window will be forced to intersect with the extent of the SpatRaster. It is envisioned that in future versions, the window may also go outside these boundaries.

# S4 method for SpatRaster
window(x)<-value

# S4 method for SpatRaster
window(x)

Arguments

x
SpatRaster

value
SpatExtent

See also

Value

none for window<- and logical for window

Examples

f <- system.file("ex/test.tif", package="terra") 
r <- rast(f)
global(r, "mean", na.rm=TRUE)

#>          mean
#> test 425.1042

e <- ext(c(179680, 180176, 331905, 332304))

window(r) <- e
global(r, "mean", na.rm=TRUE)

#>          mean
#> test 925.0588

r

#> class       : SpatRaster 
#> dimensions  : 10, 12, 1  (nrow, ncol, nlyr)
#> resolution  : 40, 40  (x, y)
#> window      : 179680, 180160, 331920, 332320  (xmin, xmax, ymin, ymax)
#> coord. ref. : +proj=sterea +lat_0=52.1561605555556 +lon_0=5.38763888888889 +k=0.9999079 +x_0=155000 +y_0=463000 +datum=WGS84 +units=m +no_defs 
#> source      : test.tif 
#> name        :  test 
#> min value   :  >138 
#> max value   : 1736< 

x <- rast(f)
xe <- crop(x, e)
global(xe, "mean", na.rm=TRUE)

#>          mean
#> test 925.0588

b <- c(xe, r)
window(b)

#> [1] FALSE  TRUE

b

#> class       : SpatRaster 
#> dimensions  : 10, 12, 2  (nrow, ncol, nlyr)
#> resolution  : 40, 40  (x, y)
#> extent (win): 179680, 180160, 331920, 332320  (xmin, xmax, ymin, ymax)
#> coord. ref. : +proj=sterea +lat_0=52.1561605555556 +lon_0=5.38763888888889 +k=0.9999079 +x_0=155000 +y_0=463000 +datum=WGS84 +units=m +no_defs 
#> sources     : memory  
#>               test.tif  
#> names       : test,  test 
#> min values  :  367,  >138 
#> max values  : 1736, 1736< 

window(r) <- NULL
r

#> class       : SpatRaster 
#> dimensions  : 115, 80, 1  (nrow, ncol, nlyr)
#> resolution  : 40, 40  (x, y)
#> extent      : 178400, 181600, 329400, 334000  (xmin, xmax, ymin, ymax)
#> coord. ref. : +proj=sterea +lat_0=52.1561605555556 +lon_0=5.38763888888889 +k=0.9999079 +x_0=155000 +y_0=463000 +datum=WGS84 +units=m +no_defs 
#> source      : test.tif 
#> name        : test 
#> min value   :  138 
#> max value   : 1736