Apply a function to the cells of two SpatRasters

Apply a function to the values of each cell of two (multilayer) SpatRasters.

# S4 method for class 'SpatRaster,SpatRaster'
xapp(x, y, fun, ..., filename="", overwrite=FALSE, wopt=list())

Arguments

x

SpatRaster

y

SpatRaster with the same geometry as x

fun

a function that operates on two vectors

...

additional arguments for fun. These are typically numerical constants. They should *never* be another SpatRaster

filename

character. Output filename

overwrite

logical. If TRUE, filename is overwritten

wopt

list with named options for writing files as in writeRaster

Value

SpatRaster

See also

Examples

r <- rast(ncols=10, nrows=10, nlyr=5)
set.seed(1)
r <- init(r, runif)
s <- init(r, runif)
x <- xapp(r, s, fun=cor)