Make tiles

Divide a SpatRaster into "tiles". The cells of another SpatRaster (normally with a much lower resolution) or a SpatVector with polygon geometry can be used to define the tiles. You can also provide one or two numbers to indicate the number of rows and columns per tile.

# S4 method for SpatRaster
makeTiles(x, y, filename="tile_.tif", extend=FALSE, na.rm=FALSE, ...)

Arguments

x

SpatRaster

y

SpatRaster, SpatVector, or numeric

filename

character. Output filename template. Filenames will be altered by adding the tile number for each tile

extend

logical. If TRUE, the extent of y is expanded to assure that it covers all of x

na.rm

logical. If TRUE, tiles with only missing values are ignored

...

additional arguments for writing files as in writeRaster

Value

character (filenames)

See also

Examples

r <- rast(ncols=100, nrows=100)
values(r) <- 1:ncell(r)
x <- rast(ncols=2, nrows=2)
filename <- paste0(tempfile(), "_.tif")
ff <- makeTiles(r, x, filename)
ff
#> [1] "/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmprLZ6yj/fileb55a5eea771e_1.tif"
#> [2] "/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmprLZ6yj/fileb55a5eea771e_2.tif"
#> [3] "/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmprLZ6yj/fileb55a5eea771e_3.tif"
#> [4] "/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmprLZ6yj/fileb55a5eea771e_4.tif"

vrt(ff)
#> class       : SpatRaster 
#> dimensions  : 100, 100, 1  (nrow, ncol, nlyr)
#> resolution  : 3.6, 1.8  (x, y)
#> extent      : -180, 180, -90, 90  (xmin, xmax, ymin, ymax)
#> coord. ref. : lon/lat WGS 84 (EPSG:4326) 
#> source      : spat_xg4kQUNIiqs8yKX_46426.vrt 
#> name        : spat_xg4kQUNIiqs8yKX_46426 
#> min value   :                          1 
#> max value   :                      10000