Merge SpatRasters, or merge a SpatVector with a data.frame

Merge multiple SpatRasters to create a new SpatRaster with a larger spatial extent. The SpatRasters must have the same origin and spatial resolution. In areas where the SpatRasters overlap, the values of the SpatRaster that is first in the sequence of arguments (or in the SpatRasterCollection) will be retained (unless first=FALSE). Cells that are NA are ignored unless na.rm=FALSE

There is also a method for merging SpatVector with a data.frame; that is, to join the data.frame to the attribute table of the SpatVector.

Usage

# S4 method for class 'SpatRaster,SpatRaster'
merge(x, y, ..., first=TRUE, na.rm=TRUE, algo=1, filename="", overwrite=FALSE, wopt=list())

# S4 method for class 'SpatRasterCollection,missing'
merge(x, first=TRUE, na.rm=TRUE, algo=1, filename="", ...)

# S4 method for class 'SpatVector,data.frame'
merge(x, y, ...)

Arguments

x

SpatRaster, SpatRasterCollection, or SpatVector

y

missing if x is a SpatRasterCollection. SpatRaster if x is a SpatRaster. data.frame if x is a SpatVector

...

if x is a SpatRaster: additional objects of the same class as x. If x is a SpatRasterCollection: options for writing files as in writeRaster. If x is a SpatVector, the same arguments as in merge

first

logical. If TRUE, in areas where rasters overlap, the first value is used. Otherwise the last value is used

na.rm

logical. If TRUE values that are NA are ignored

algo

integer. You can use 1, 2, or 3 to pick a merge algorithm. 1 is generally faster than 2, but file compression may be poor. 2 is slower but has better compression. In algo 3 a vrt)) is used and returned if no filename is specified; otherwise its performance is like that of algo 2 {filename}{character. Output filename} {overwrite}{logical. If TRUE, filename is overwritten} {wopt}{list with named options for writing files as in writeRaster}

SpatRaster or SpatVector

Combining tiles with vrt may be more efficient than using merge. See mosaic for averaging overlapping regions. See classify to merge a SpatRaster and a data.frame and union to combine SpatExtent objects.

x <- rast(xmin=-110, xmax=-80, ymin=40, ymax=70, res=1, vals=1) y <- rast(xmin=-85, xmax=-55, ymax=60, ymin=30, res=1, vals=2) z <- rast(xmin=-60, xmax=-30, ymax=50, ymin=20, res=1, vals=3)m1 <- merge(x, y, z) m2 <- merge(z, y, x) m3 <- merge(y, x, z)# if you have many SpatRasters, it may be convenient # to make a SpatRasterCollection # s <- sprc(list(x, y, z)) s <- sprc(x, y, z)sm1 <- merge(s, algo=1, first=FALSE) sm2 <- merge(s, algo=2, first=FALSE) sm3 <- merge(s, algo=3, first=FALSE) # panel(c(sm1, sm2, sm3))## SpatVector with data.frame f <- system.file("ex/lux.shp", package="terra") p <- vect(f) dfr <- data.frame(District=p$NAME_1, Canton=p$NAME_2, Value=round(runif(length(p), 100, 1000))) dfr <- dfr[1:5, ] pm <- merge(p, dfr, all.x=TRUE, by.x=c('NAME_1', 'NAME_2'), by.y=c('District', 'Canton')) pm values(pm)

methods spatial