Merge SpatRaster or SpatExtent objects, or a SpatVector with a data.frame

Merge SpatRasters to form a new SpatRaster object with a larger spatial extent. If objects overlap, the values get priority in the same order as the arguments. The SpatRasters must have the same origin and spatial resolution. In areas where the SpatRaster objects overlap, the values of the SpatRaster that is last in the sequence of arguments will be retained. See classify to merge a SpatRaster and a data.frame. You can also merge SpatExtent objects.

There is a also a method for merging SpatVector with a data.frame; that is, to join the data.frame to the attribute table of the SpatVector.

# S4 method for SpatRaster,SpatRaster
merge(x, y, ..., filename="", overwrite=FALSE, wopt=list())

# S4 method for SpatRasterCollection,missing
merge(x, filename="", ...)

# S4 method for SpatExtent,SpatExtent
merge(x, y, ...)

# S4 method for SpatVector,data.frame
merge(x, y, ...)

Arguments

x

SpatRaster or SpatExtent

y

object of same class as x

...

if x is a SpatRaster: additional objects of the same class as x. If x is a SpatRasterCollection: options for writing files as in writeRaster. If x is a SpatVector, the same arguments as in merge

filename

character. Output filename

overwrite

logical. If TRUE, filename is overwritten

wopt

list with named options for writing files as in writeRaster

Value

SpatRaster or SpatExtent

See also

Note

You can use merge with do.call to merge a list of SpatRasters (see example). But note that if the list is named, these names are used by merge. So if all elements are named, there should be one element with a SpatRaster called x and another one called y. For example with names(x)[1:2] <- c("x"m "y"). You can also removed the names of the the first two elements (assuming these are SpatRasters) with names(x)[1:2] <- "".

Examples

x <- rast(xmin=-110, xmax=-80, ymin=40, ymax=70, ncols=30, nrows=30)
y <- rast(xmin=-85, xmax=-55, ymax=60, ymin=30, ncols=30, nrows=30)
z <- rast(xmin=-60, xmax=-30, ymax=50, ymin=20, ncols=30, nrows=30)
values(x) <- 1:ncell(x)
values(y) <- 1:ncell(y)
values(z) <- 1:ncell(z)

m1 <- merge(x, y, z)
m2 <- merge(z, y, x)
m3 <- merge(y, x, z)

# if you have many SpatRasters make a SpatRasterCollection from a list
rlist <- list(x, y, z)
rsrc <- sprc(rlist)

m <- merge(rsrc)

## SpatVector with data.frame
f <- system.file("ex/lux.shp", package="terra")
p <- vect(f)
dfr <- data.frame(District=p$NAME_1, Canton=p$NAME_2, Value=round(runif(length(p), 100, 1000)))
dfr <- dfr[1:5, ]
pm <- merge(p, dfr, all.x=TRUE, by.x=c('NAME_1', 'NAME_2'), by.y=c('District', 'Canton'))
pm
#>  class       : SpatVector 
#>  geometry    : polygons 
#>  dimensions  : 12, 7  (geometries, attributes)
#>  extent      : 5.74414, 6.528252, 49.44781, 50.18162  (xmin, xmax, ymin, ymax)
#>  coord. ref. : lon/lat WGS 84 (EPSG:4326) 
#>  names       :   NAME_1   NAME_2  ID_1  ID_2  AREA   POP Value
#>  type        :    <chr>    <chr> <num> <num> <num> <int> <num>
#>  values      : Diekirch Clervaux     1     1   312 18081   708
#>                Diekirch Diekirch     1     2   218 32543   755
#>                Diekirch  Redange     1     3   259 18664   568
values(pm)
#>          NAME_1           NAME_2 ID_1 ID_2 AREA    POP Value
#> 1      Diekirch         Clervaux    1    1  312  18081   708
#> 2      Diekirch         Diekirch    1    2  218  32543   755
#> 3      Diekirch          Redange    1    3  259  18664   568
#> 4      Diekirch          Vianden    1    4   76   5163   878
#> 5      Diekirch            Wiltz    1    5  263  16735   496
#> 6  Grevenmacher       Echternach    2    6  188  18899    NA
#> 7  Grevenmacher           Remich    2    7  129  22366    NA
#> 8  Grevenmacher     Grevenmacher    2   12  210  29828    NA
#> 9    Luxembourg         Capellen    3    8  185  48187    NA
#> 10   Luxembourg Esch-sur-Alzette    3    9  251 176820    NA
#> 11   Luxembourg       Luxembourg    3   10  237 182607    NA
#> 12   Luxembourg           Mersch    3   11  233  32112    NA