mosaic SpatRasters

mosaic SpatRasters to form a new SpatRaster object with a larger spatial extent. Unlike with merge, values on overlapping areas are averaged.

# S4 method for SpatRaster,SpatRaster
mosaic(x, y, ..., fun="mean", filename="", overwrite=FALSE, wopt=list())

# S4 method for SpatRasterCollection,missing
mosaic(x, fun="mean", filename="", ...)

Arguments

x
SpatRaster

y
object of same class as x

...
additional SpatRasters

fun
character. One of "sum", "mean", "median", "min", "max"

filename
character. Output filename

overwrite
logical. If TRUE, filename is overwritten

wopt
list with named options for writing files as in writeRaster

Details

The SpatRaster objects must have the same origin and spatial resolution. In areas where the SpatRaster objects overlap, the values of the SpatRaster that is last in the sequence of arguments will be retained.

Value

SpatRaster

See also

Examples

x <- rast(xmin=-110, xmax=-80, ymin=40, ymax=70, ncols=30, nrows=30)
y <- rast(xmin=-85, xmax=-55, ymax=60, ymin=30, ncols=30, nrows=30)
z <- rast(xmin=-60, xmax=-30, ymax=50, ymin=20, ncols=30, nrows=30)
values(x) <- 1:ncell(x)
values(y) <- 1:ncell(y)
values(z) <- 1:ncell(z)

m1 <- mosaic(x, y, z)
m2 <- mosaic(z, y, x)

# if you have many SpatRasters make a SpatRasterCollection from a list
rlist <- list(x, y, z)
rsrc <- src(rlist)

m <- mosaic(rsrc)