Create a SpatRasterDataset

Methods to create a SpatRasterDataset. This is an object to hold "sub-datasets", each a SpatRaster that in most cases will have multiple layers.

See describe for getting information about the sub-datasets present in a file.

# S4 method for missing
sds(x) 

# S4 method for character
sds(x, ids=0)

# S4 method for SpatRaster
sds(x, ...) 

# S4 method for list
sds(x)

Arguments

x
character (filename), or SpatRaster, or list of SpatRaster objects, or missing. If multiple filenames are provided, it is attempted to make SpatRasters from these, and combine them into a SpatRasterDataset

ids
optional. vector of integer subdataset ids. Ignored if the first value is not a positive integer

...
additional arguments. Can be other SpatRaster objects if x is a SpatRaster

Value

SpatRasterDataset

See also

Examples

s <- rast(system.file("ex/logo.tif", package="terra"))   
x <- sds(s, s/2)
names(x) <- c("first", "second")
x

#> class       : SpatRasterDataset 
#> subdatasets : 2 
#> dimensions  : 77, 101 (nrow, ncol)
#> nlyr        : 3, 3 
#> resolution  : 1, 1  (x, y)
#> extent      : 0, 101, 0, 77  (xmin, xmax, ymin, ymax)
#> coord. ref. : +proj=merc +lon_0=0 +k=1 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs 
#> source(s)   : logo.tif, memory 
#> names       : first, second 

length(x)

#> [1] 2

# extract the second SpatRaster
x[2]

#> class       : SpatRaster 
#> dimensions  : 77, 101, 3  (nrow, ncol, nlyr)
#> resolution  : 1, 1  (x, y)
#> extent      : 0, 101, 0, 77  (xmin, xmax, ymin, ymax)
#> coord. ref. : +proj=merc +lon_0=0 +k=1 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs 
#> source      : memory 
#> names       :   red, green,  blue 
#> min values  :     0,     0,     0 
#> max values  : 127.5, 127.5, 127.5