Cell values and geometry attributes

Get the cell values of a SpatRaster or the attributes of a SpatVector

# S4 method for SpatRaster
values(x, mat=TRUE, dataframe=FALSE, row=1, nrows=nrow(x), col=1, ncols=ncol(x))

# S4 method for SpatVector
values(x)

Arguments

x
SpatRaster or SpatVector

mat
logical. If TRUE, values are returned as a matrix instead of as a vector, except when dataframe is TRUE

dataframe
logical. If TRUE, values are returned as a data.frame instead of as a vector (also if matrix is TRUE)

row
positive integer. Row number to start from, should be between 1 and nrow(x)

nrows
positive integer. How many rows?

col
positive integer. Column number to start from, should be between 1 and ncol(x)

ncols
positive integer. How many columns? Default is the number of columns left after the start column

Details

If x is a SpatRaster, and matrix=TRUE, a matrix is returned in which the values of each layer are represented by a column (with ncell(x) rows). The values per layer are in cell-order, that is, from top-left, to top-right and then down by row. Use as.matrix for an alternative matrix representation where the number of rows and columns matches that of x, if x has a single layer. If matrix=FALSE, the values are returned as a vector. In cell-order by layer.

Value

matrix or data.frame

See also

Examples

r <- rast(system.file("ex/test.tif", package="terra"))
r

#> class       : SpatRaster 
#> dimensions  : 115, 80, 1  (nrow, ncol, nlyr)
#> resolution  : 40, 40  (x, y)
#> extent      : 178400, 181600, 329400, 334000  (xmin, xmax, ymin, ymax)
#> coord. ref. : +proj=sterea +lat_0=52.1561605555556 +lon_0=5.38763888888889 +k=0.9999079 +x_0=155000 +y_0=463000 +datum=WGS84 +units=m +no_defs 
#> source      : test.tif 
#> name        : test 
#> min value   :  138 
#> max value   : 1736 

v <- values(r)
head(v)

#>      test
#> [1,]  NaN
#> [2,]  NaN
#> [3,]  NaN
#> [4,]  NaN
#> [5,]  NaN
#> [6,]  NaN

f <- system.file("ex/lux.shp", package="terra")
v <- vect(f)
x <- values(v)
head(x)

#>   ID_1       NAME_1 ID_2     NAME_2 AREA
#> 1    1     Diekirch    1   Clervaux  312
#> 2    1     Diekirch    2   Diekirch  218
#> 3    1     Diekirch    3    Redange  259
#> 4    1     Diekirch    4    Vianden   76
#> 5    1     Diekirch    5      Wiltz  263
#> 6    2 Grevenmacher    6 Echternach  188