Zonal statistics

Compute zonal statistics, that is summarize values of a SpatRaster for each "zone" defined by another SpatRaster.

If fun is a true R function, zonal may fail for very large SpatRaster objects, except for the functions ("mean", "min", "max", "sum", "isNA", and "notNA").

You can also summarize values of a SpatVector for each polygon (zone) defined by another SpatVector.

See extract to summarize values of a SpatRaster with a SpatVector.

# S4 method for SpatRaster,SpatRaster
zonal(x, z, fun=mean, ..., as.raster=FALSE, filename="", wopt=list()) 

# S4 method for SpatVector,SpatVector
zonal(x, z, fun=mean, ..., weighted=FALSE, as.polygons=FALSE)

Arguments

x

SpatRaster or SpatVector

z

Object of the same class as x. That is, either a SpatRaster with cell-values representing zones or a SpatVector with each polygon geometry representing a zone

fun

function to be applied to summarize the values by zone. Either as character: "mean", "min", "max", "sum", "isNA", and "notNA" and, for relatively small SpatRasters, a proper function

...

additional arguments passed to fun

as.raster

logical. If TRUE, a SpatRaster is returned with the zonal statistic for each zone

filename

character. Output filename (ignored if as.raster=FALSE

wopt

list with additional arguments for writing files as in writeRaster

weighted

logical. If TRUE, a weighted.mean is computed and fun is ignored. Weights are based on the length of the lines or the area of the polygons in x that intersect with z. This argument is ignored of x is a SpatVector or points

as.polygons

logical. Should the results be merged with the attributes of z?

Value

A data.frame with a value for each zone) or a SpatRaster or SpatVector of polygons

See also

Examples

r <- rast(ncols=10, nrows=10)
values(r) <- 1:ncell(r)
z <- rast(r)
values(z) <- rep(c(1:2, NA, 3:4), each=20)
names(z) <- "zone"
zonal(r, z, "sum", na.rm=TRUE)
#>   zone lyr.1
#> 1    1   210
#> 2    2   610
#> 3    3  1410
#> 4    4  1810

# multiple layers
r <- rast(system.file("ex/logo.tif", package = "terra")) 
# zonal layer 
z <- rast(r, 1)
names(z) <- "zone"
values(z) <- rep(c(1:2, NA, c(3:4)), each=ncell(r)/5, length.out=ncell(r))

zonal(r, z, "mean", na.rm = TRUE)
#>   zone      red    green     blue
#> 1    1 197.9486 198.0103 193.5556
#> 2    2 173.2219 176.7717 185.2585
#> 3    3 168.2952 172.6232 184.6939
#> 4    4 193.5859 197.0019 206.5717

# raster of zonal values
zr <- zonal(r, z, "mean", na.rm = TRUE, as.raster=TRUE)

### SpatVector 

f <- system.file("ex/lux.shp", package="terra")
v <- vect(f)[,c(2,4)]

p <- spatSample(v, 100)
values(p) <- data.frame(b2=1:100, ssep1=100:1)

zonal(p, v, mean)
#>    zone       b2    ssep1
#> 1     1 66.33333 34.66667
#> 2     2 69.40000 31.60000
#> 3     3 39.75000 61.25000
#> 4     4 41.00000 60.00000
#> 5     5 60.80000 40.20000
#> 6     6 44.28571 56.71429
#> 7     7 61.71429 39.28571
#> 8     8 64.50000 36.50000
#> 9     9 34.42857 66.57143
#> 10   10 47.75000 53.25000
#> 11   11 39.11111 61.88889
#> 12   12 58.21429 42.78571